services/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
feedparser==6.0.11
python-dateutil==2.9.0.post0
redis==5.0.8
